parallel pairwise dist metric

+++ analytics.py
@@ -139,7 +139,7 @@ if __name__ == "__main__":
     data = load_flat("flat.csv")
     keys = list(data.keys())
     P = np.array([data[k][0] for k in keys]).astype(float)
-    M = np.array([data[k][1] for k in keys]).astype(int)
+    M = np.array([data[k][1] for k in keys]).astype(float)
 
     print("Computing pairwise distance matrix")
     dim = (192, 108)
